Refactor
* Reorganised profile project structure
