+++ profile/src/main/java/com/scottlogic/deg/profile/ProfileSchemaFileLoader.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.scottlogic.deg.profile.dtos;
+package com.scottlogic.deg.profile;
 
 import com.scottlogic.deg.common.ValidationException;
