+++ profile/src/main/java/com/scottlogic/deg/profile/SchemaVersionGetter.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.scottlogic.deg.profile.serialisation;
+package com.scottlogic.deg.profile;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.scottlogic.deg.profile.dtos.SchemaDto;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
