+++ profile/src/main/java/com/scottlogic/deg/profile/SchemaVersionValidator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.scottlogic.deg.profile.dtos;
+package com.scottlogic.deg.profile;
 
 import java.io.IOException;
 import java.net.URL;
