+++ profile/src/main/java/com/scottlogic/deg/profile/dtos/constraints/ConstraintDTO.java
@@ -18,7 +18,7 @@ package com.scottlogic.deg.profile.dtos.constraints;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.scottlogic.deg.profile.ConstraintDeserializer;
+import com.scottlogic.deg.profile.serialisation.ConstraintDeserializer;
 import com.scottlogic.deg.profile.common.ConstraintType;
 
 @JsonDeserialize(using = ConstraintDeserializer.class)
