+++ profile/src/main/java/com/scottlogic/deg/profile/serialisation/ConstraintDeserializer.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.profile;
+package com.scottlogic.deg.profile.serialisation;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
