+++ profile/src/test/java/com/scottlogic/deg/profile/reader/file/names/NameRetrieverTest.java
@@ -19,6 +19,7 @@ package com.scottlogic.deg.profile.reader.file.names;
 
 import com.scottlogic.deg.generator.profile.constraints.atomic.NameConstraintTypes;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedList;
+import com.scottlogic.deg.profile.reader.NameRetriever;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
